requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
Pillow==7.1.2
pycryptodome==3.9.7
requests==2.23.0
urllib3==1.25.9
